requirements.txt
fastapi
uvicorn[standard]
python-multipart
groq
google-genai
openai
anthropic
